Added wifi time functionality and split WifiClock.cpp into many files.
