+++ arduino/WifiClock/src/BoardConstants.h
@@ -1,0 +1,25 @@
+#ifndef BOARDCONSTANTS_H
+#define BOARDCONSTANTS_H
+
+#define RLED	0
+#define GLED	4
+#define BLED	5
+
+#define BEEPER	16
+
+#define MODE	9
+#define INCR	10
+#define DECR	2
+
+#define CLK		14
+#define DIN		13
+#define LOAD	15
+
+#define DIG0	0
+#define DIG1	1
+#define DIG2	2
+#define DIG3	3
+#define DIG4	4
+#define DIG5	5
+
+#endif
