+++ arduino/WifiClock/src/NoteConstants.h
@@ -1,0 +1,58 @@
+#ifndef NOTECONSTANTS_H
+#define NOTECONSTANTS_H
+
+#define NOTE_C3		130.8
+#define NOTE_CS3	138.6
+#define NOTE_DF3	138.6
+#define NOTE_D3		146.8
+#define NOTE_DS3	155.6
+#define NOTE_EF3	155.6
+#define NOTE_E3		164.8
+#define NOTE_F3		174.6
+#define NOTE_FS3	185.0
+#define NOTE_GF3	185.0
+#define NOTE_G3		196.0
+#define NOTE_GS3	207.7
+#define NOTE_AF3	207.7
+#define NOTE_A3		220.0
+#define NOTE_AS3	233.1
+#define NOTE_BF3	233.1
+#define NOTE_B3		246.9
+
+#define NOTE_C4		261.6
+#define NOTE_CS4	277.2
+#define NOTE_DF4	277.2
+#define NOTE_D4		293.7
+#define NOTE_DS4	311.1
+#define NOTE_EF4	311.1
+#define NOTE_E4		329.6
+#define NOTE_F4		349.2
+#define NOTE_FS4	370.0
+#define NOTE_GF4	370.0
+#define NOTE_G4		392.0
+#define NOTE_GS4	415.3
+#define NOTE_AF4	415.3
+#define NOTE_A4		440.0
+#define NOTE_AS4	466.2
+#define NOTE_BF4	466.2
+#define NOTE_B4		493.9
+
+#define NOTE_C5		523.3
+#define NOTE_CS5	554.4
+#define NOTE_DF5	554.4
+#define NOTE_D5		587.3
+#define NOTE_DS5	622.3
+#define NOTE_EF5	622.3
+#define NOTE_E5		659.3
+#define NOTE_F5		698.5
+#define NOTE_FS5	740.0
+#define NOTE_GF5	740.0
+#define NOTE_G5		784.0
+#define NOTE_GS5	830.6
+#define NOTE_AF5	830.6
+#define NOTE_A5		880.0
+#define NOTE_AS5	932.3
+#define NOTE_BF5	932.3
+#define NOTE_B5		987.8
+
+#endif
