+++ arduino/WifiClock/src/WifiClock.h
@@ -1,87 +1,25 @@
 #ifndef WIFICLOCK_H
 #define WIFICLOCK_H
 
+// built in header files
 #include <unordered_map>
 #include "Arduino.h"
-#include "LedControl.h"
-#include "Clock.h"
+
+// ESP8266 header files
+#include "ESP8266WiFi.h"
+#include "WiFiUdp.h"
 #include "Ticker.h"
 
-#define RLED	0
-#define GLED	4
-#define BLED	5
+// dependency header files
+#include "LedControl.h"
+#include "NTPClient.h"
 
-#define BEEPER	16
+// project header files
+#include "NoteConstants.h"
+#include "BoardConstants.h"
+#include "Clock.h"
 
-#define MODE	9
-#define INCR	10
-#define DECR	2
-
-#define CLK		14
-#define DIN		13
-#define LOAD	15
-
-#define DIG0	0
-#define DIG1	1
-#define DIG2	2
-#define DIG3	3
-#define DIG4	4
-#define DIG5	5
-
-#define NOTE_C3		130.8
-#define NOTE_CS3	138.6
-#define NOTE_DF3	138.6
-#define NOTE_D3		146.8
-#define NOTE_DS3	155.6
-#define NOTE_EF3	155.6
-#define NOTE_E3		164.8
-#define NOTE_F3		174.6
-#define NOTE_FS3	185.0
-#define NOTE_GF3	185.0
-#define NOTE_G3		196.0
-#define NOTE_GS3	207.7
-#define NOTE_AF3	207.7
-#define NOTE_A3		220.0
-#define NOTE_AS3	233.1
-#define NOTE_BF3	233.1
-#define NOTE_B3		246.9
-
-#define NOTE_C4		261.6
-#define NOTE_CS4	277.2
-#define NOTE_DF4	277.2
-#define NOTE_D4		293.7
-#define NOTE_DS4	311.1
-#define NOTE_EF4	311.1
-#define NOTE_E4		329.6
-#define NOTE_F4		349.2
-#define NOTE_FS4	370.0
-#define NOTE_GF4	370.0
-#define NOTE_G4		392.0
-#define NOTE_GS4	415.3
-#define NOTE_AF4	415.3
-#define NOTE_A4		440.0
-#define NOTE_AS4	466.2
-#define NOTE_BF4	466.2
-#define NOTE_B4		493.9
-
-#define NOTE_C5		523.3
-#define NOTE_CS5	554.4
-#define NOTE_DF5	554.4
-#define NOTE_D5		587.3
-#define NOTE_DS5	622.3
-#define NOTE_EF5	622.3
-#define NOTE_E5		659.3
-#define NOTE_F5		698.5
-#define NOTE_FS5	740.0
-#define NOTE_GF5	740.0
-#define NOTE_G5		784.0
-#define NOTE_GS5	830.6
-#define NOTE_AF5	830.6
-#define NOTE_A5		880.0
-#define NOTE_AS5	932.3
-#define NOTE_BF5	932.3
-#define NOTE_B5		987.8
-
+// constants
 #define MAX_INT		999999
 #define MIN_INT		-99999
 #define MAX_HEX		0xFFFFFF
@@ -132,9 +70,6 @@ class WifiClock
     static void decr_button_callback(void (*func)(void), int mode);
 	static byte debounce(int button);
 	
-	// Timer Functions
-	void timer_callback(float period, void (*func)(void));
-	
 	// Beeper Functions
 	static void play_note(float frequency);
 	static void stop_note(void);
@@ -158,19 +93,24 @@ class WifiClock
 	bool event_scheduled(Time t);
 	bool event_scheduled(void);
 	
-	// Not Yet Implemented
-	/*void connect_to_wifi(char* ssid, char* password);
-    void set_time(int hour, int minute, int second);
-    void set_time_ntp(int utc_offset);
-    void start_clock24(void);*/
-	//void clock_timer_callback(void (*func)(void));
-	//void _update_clock24(void);
+	// Wifi Time Functions
+	void connect_to_wifi(const char* ssid, const char* password);
+	bool check_connection(void);
+	void start_wifi_time(const char* ssid, const char* password);
+	void start_wifi_time(const char* ssid, const char* password, short offset);
+	void stop_wifi_time(void);
+	void set_time_offset(short offset);
+	void set_military_time(bool set);
+	void update_time_from_wifi(void);
+	unsigned short _correct_hours(void);
+	void display_time(bool display_pm_light=true, int pin=RLED);
+	void update_and_display_time(bool display_pm_light=true, int pin=RLED);
+	
 	
   private:
     const int _digits[6] = {DIG0, DIG1, DIG2, DIG3, DIG4, DIG5};
 	LedControl _lc = LedControl(DIN, CLK, LOAD, 1);					// Initialize MAX7219
 	Schedule _schedule;												// Initialize unordered map
-	Ticker _timer;													// Initialize timer
 	Clock _clock = Clock();											// Initialize a clock
 	byte _curr_type = 0;											// 0: nothing
 																	// 1: int
@@ -182,6 +122,15 @@ class WifiClock
 	byte _curr_dp = 0x0;
 	bool _curr_right = true;
 	bool _curr_zeros = true;
+	
+	WiFiUDP _ntpUDP;
+	NTPClient _timeClient = NTPClient(_ntpUDP);
+	unsigned short _hours;
+	unsigned short _minutes;
+	unsigned short _seconds;
+	bool _pm;
+	short _offset = 0;
+	bool _military_time = true;
 	
 	// Display Helper Functions
 	int _count(int n);
